refactor: replace inline styles with Tailwind classes using twMerge
