+++ src/features/select-character/card/Card.tsx
@@ -1,4 +1,5 @@
 import { type ChangeEvent, type FC } from 'react';
+import { twMerge } from 'tailwind-merge';
 
 import { useAppDispatch, useAppSelector } from '@/state/store/hooks.ts';
 import { toggleItemSelection } from '@/state/store/selectedItemsSlice.ts';
@@ -37,19 +38,24 @@ export const Card: FC<CardProps> = ({
       })
     );
   };
+
+  const statusColor =
+    status === 'Alive'
+      ? 'bg-[hsl(140,70%,45%)]'
+      : status === 'Dead'
+        ? 'bg-[hsl(0,70%,50%)]'
+        : 'bg-[hsl(0,0%,50%)]';
+
   return (
     <article
-      className="rounded-xl overflow-hidden shadow-lg transition-all duration-300 w-full max-w-xs h-[420px] flex flex-col border"
-      style={{
-        backgroundColor: isSelected
-          ? 'var(--color-bg-secondary)'
-          : 'var(--color-bg)',
-        borderColor: isSelected ? 'var(--color-accent)' : 'var(--color-border)',
-      }}
+      className={twMerge(
+        'rounded-xl overflow-hidden shadow-lg transition-all duration-300 w-full max-w-xs h-[420px] flex flex-col border',
+        isSelected ? 'bg-bg-secondary border-accent' : 'bg-bg border-border'
+      )}
     >
       <div className="relative h-64 flex-shrink-0">
         <img src={image} alt={name} className="w-full h-full object-cover" />
-        <div className="absolute top-2 left-2 ">
+        <div className="absolute top-2 left-2">
           <input
             type="checkbox"
             checked={isChecked}
@@ -59,46 +65,31 @@ export const Card: FC<CardProps> = ({
         </div>
       </div>
       <section
-        className="p-4 flex-1 flex flex-col"
-        style={{
-          background: isSelected
-            ? 'linear-gradient(to bottom, var(--color-bg-secondary), var(--color-bg))'
-            : undefined,
-        }}
+        className={twMerge(
+          'p-4 flex-1 flex flex-col',
+          isSelected && 'bg-gradient-to-b from-bg-secondary to-bg'
+        )}
       >
         <h3
-          className="text-xl font-bold mb-2 truncate"
-          style={{
-            color: isSelected ? 'var(--color-accent)' : 'var(--text-primary)',
-          }}
+          className={twMerge(
+            'text-xl font-bold mb-2 truncate',
+            isSelected ? 'text-accent' : 'text-text-primary'
+          )}
         >
           {name}
         </h3>
-        <p
-          className="mb-1 flex items-center"
-          style={{ color: 'var(--color-text-secondary)' }}
-        >
+        <p className="mb-1 flex items-center text-text-secondary">
           <span
-            className="inline-block w-2 h-2 rounded-full mr-2 flex-shrink-0"
-            style={{
-              backgroundColor:
-                status === 'Alive'
-                  ? 'hsl(140, 70%, 45%)'
-                  : status === 'Dead'
-                    ? 'hsl(0, 70%, 50%)'
-                    : 'hsl(0, 0%, 50%)',
-            }}
+            className={twMerge(
+              'inline-block w-2 h-2 rounded-full mr-2 flex-shrink-0',
+              statusColor
+            )}
           ></span>
           <span className="truncate">
             {status} — {species}
           </span>
         </p>
-        <p
-          className="truncate"
-          style={{ color: 'var(--color-text-secondary)' }}
-        >
-          {gender}
-        </p>
+        <p className="truncate text-text-secondary">{gender}</p>
       </section>
     </article>
   );
